Fix OSPhysicalToCached etc for 64-bit pointers

+++ include/dolphin/os.h
@@ -206,8 +206,8 @@ void* OSCachedToUncached(void* caddr);
 void* OSUncachedToCached(void* ucaddr);
 
 #if !DEBUG
-#define OSPhysicalToCached(paddr)    ((void*) ((u32)(OS_BASE_CACHED   + (u32)(paddr))))
-#define OSPhysicalToUncached(paddr)  ((void*) ((u32)(OS_BASE_UNCACHED + (u32)(paddr))))
+#define OSPhysicalToCached(paddr)    ((void*) ((uintptr_t)(OS_BASE_CACHED   + (uintptr_t)(paddr))))
+#define OSPhysicalToUncached(paddr)  ((void*) ((uintptr_t)(OS_BASE_UNCACHED + (uintptr_t)(paddr))))
 #define OSCachedToPhysical(caddr)    ((u32)   ((u32)(caddr)  - OS_BASE_CACHED))
 #define OSUncachedToPhysical(ucaddr) ((u32)   ((u32)(ucaddr) - OS_BASE_UNCACHED))
 #define OSCachedToUncached(caddr)    ((void*) ((u8*)(caddr)  + (OS_BASE_UNCACHED - OS_BASE_CACHED)))
